ClickHouseDriver.Core.Client

Description

This module provides implementations of user's APIs

Data Fetch and Insert

query

Arguments

:: Env () w
Haxl environment for connection

-> String
Query command for SELECT and SHOW only

-> IO (Either String (Vector (Vector ClickhouseType)))

query command

queryWithInfo :: String -> Env () w -> IO (Either String CKResult)

query result contains query information.

deploySettings :: TCPConnection -> IO (Env () w)

insertMany :: Env () w -> String -> [[ClickhouseType]] -> IO ByteString

insertOneRow

Arguments

:: Env () w

-> String
SQL command

-> [ClickhouseType]
a row of local clickhouse data type to be serialize and insert.

-> IO ByteString
The result bytestring indicate success or failure.

ping :: Env () w -> IO ()

ping pong

withQuery

Arguments

:: Env () w
enviroment i.e. the database resource

-> String
sql statement

-> (Either String (Vector (Vector ClickhouseType)) -> IO a)
callback function that returns type a

-> IO a
type a wrapped in IO monad.

fetch

Arguments

:: String
SQL SELECT command

-> GenHaxl u w (Either String (Vector (Vector ClickhouseType)))
result wrapped in Haxl monad for other tasks run with concurrency.

fetch data only

fetchWithInfo :: String -> GenHaxl u w (Either String CKResult)

fetch data alone with query information

execute :: Env u w -> GenHaxl u w a -> IO a

for general use e.g. creating table, | multiple queries, multiple insertions.

Communication

client

Arguments

:: Resource a

=> Either String a

-> IO (Env () w)
Either wrong message of resource with type a

defaultClient :: IO (Env () w)

closeClient :: Env () w -> IO ()

close connection

Connection pool

defaultClientPool

Arguments

:: Int
number of stripes

-> NominalDiffTime
idle time for each stripe

-> Int
maximum resources for reach stripe

-> IO (Env () w)
Haxl env wrapped in IO monad.

createClient :: ConnParams -> IO (Env () w)

createClientPool

Arguments

:: ConnParams
parameters for connection settings

-> Int
number of stripes

-> NominalDiffTime
idle time for each stripe

-> Int
maximum resources for reach stripe

-> IO (Env () w)