ClickHouseDriver.Core.Column

Description

This module contains the implementations of serialization and deserialization of Clickhouse data types.

Serialize and deserialize

data ClickhouseType

Constructors

CKInt8 Int8

CKInt16 Int16

CKInt32 Int32

CKInt64 Int64

CKInt128 Int64 Int64

CKUInt8 Word8

CKUInt16 Word16

CKUInt32 Word32

CKUInt64 Word64

CKUInt128 Word64 Word64

CKString ByteString

CKTuple (Vector ClickhouseType)

CKArray (Vector ClickhouseType)

CKDecimal Float

CKDecimal32 Float

CKDecimal64 Double

CKDecimal128 Double

CKIPv4 (Word8, Word8, Word8, Word8)

CKIPv6 (Word16, Word16, Word16, Word16, Word16, Word16, Word16, Word16)

CKDate

Fields

year :: !Integer

month :: !Int

day :: !Int

CKNull

readColumn

Arguments

:: ServerInfo
Server information is needed in case of some parameters are missing

-> Int
number of rows

-> ByteString
data type

-> Reader (Vector ClickhouseType)

writeColumn

Arguments

:: Context
context contains client information and server information

-> ByteString
column name

-> ByteString
column type (String, Int, etc)

-> Vector ClickhouseType
items to be serialized.

-> Writer Builder
result wrapped in a customized Writer Monad used for concatenating string builders.

Operations on ClickhouseType

transpose :: Vector (Vector ClickhouseType) -> Vector (Vector ClickhouseType)

putStrLn :: Vector (Vector ClickhouseType) -> IO ()

print in format