ClickHouseDriver.Core.HTTP

Documentation

data HttpConnection

Constructors

HttpConnection

Fields

httpParams :: !HttpParams

basic parameters

httpManager :: !Manager

http manager

data HttpParams

Constructors

HttpParams

Fields

httpHost :: !String

httpPort :: !Int

httpUsername :: !String

httpPassword :: !String

httpDatabase :: Maybe String

data Format

Constructors

CSV

JSON

TUPLE

type Haxl a = GenHaxl () a

type Cmd = String

type JSONResult = Either ByteString [HashMap Text Value]

getByteString :: String -> GenHaxl u w ByteString

Fetch data from ClickHouse client in the text format.

getText :: String -> GenHaxl u w Text

getJSON :: String -> GenHaxl u w JSONResult

Fetch data from ClickHouse client in the JSON format.

getTextM :: (Monad m, Traversable m) => m String -> GenHaxl u w (m Text)

Fetch data from Clickhouse client with commands warped in a Traversable monad.

getJsonM :: (Monad m, Traversable m) => m String -> GenHaxl u w (m JSONResult)

Fetch data from Clickhouse client in the format of JSON

exec :: HttpEnvironment a => String -> Env a w -> IO (Either ByteString String)

actual function used by user to perform fetching command

insertOneRow :: HttpEnvironment a => String -> [ClickhouseType] -> Env a w -> IO (Either ByteString String)

insert one row

insertMany :: HttpEnvironment a => String -> [[ClickhouseType]] -> Env a w -> IO (Either ByteString String)

insert one or more rows

insertFromFile :: HttpEnvironment a => String -> Format -> FilePath -> Env a w -> IO (Either ByteString String)

insert data from

ping :: GenHaxl u w ByteString

setupEnv :: (MonadIO m, HttpEnvironment a) => a -> m (Env a w)

Default environment

defaultHttpClient :: MonadIO m => m (Env HttpConnection w)

defaultHttpPool :: MonadIO m => Int -> NominalDiffTime -> Int -> m (Env (Pool HttpConnection) w)

httpClient :: MonadIO m => String -> String -> m (Env HttpConnection w)

runQuery :: MonadIO m => Env u w -> GenHaxl u w a -> m a

rename runHaxl function.

data HttpConnection

Constructors

HttpConnection

Fields

httpParams :: !HttpParams

basic parameters

httpManager :: !Manager

http manager

defaultHttpConnection :: IO HttpConnection

createHttpPool :: HttpParams -> Int -> NominalDiffTime -> Int -> IO (Pool HttpConnection)

httpConnect :: String -> String -> Int -> String -> IO HttpConnection

httpConnectDb :: String -> String -> Int -> String -> Maybe String -> IO HttpConnection