ClickHouseDriver.Core.HTTP.Client

Description

This module provides implementation of user's API

Setting

setupEnv :: (MonadIO m, HttpEnvironment a) => a -> m (Env a w)

Default environment

runQuery :: MonadIO m => Env u w -> GenHaxl u w a -> m a

rename runHaxl function.

Query

getByteString :: String -> GenHaxl u w ByteString

Fetch data from ClickHouse client in the text format.

getJSON :: String -> GenHaxl u w JSONResult

Fetch data from ClickHouse client in the JSON format.

getText :: String -> GenHaxl u w Text

getTextM :: (Monad m, Traversable m) => m String -> GenHaxl u w (m Text)

Fetch data from Clickhouse client with commands warped in a Traversable monad.

getJsonM :: (Monad m, Traversable m) => m String -> GenHaxl u w (m JSONResult)

Fetch data from Clickhouse client in the format of JSON

insertOneRow :: HttpEnvironment a => String -> [ClickhouseType] -> Env a w -> IO (Either ByteString String)

insert one row

insertMany :: HttpEnvironment a => String -> [[ClickhouseType]] -> Env a w -> IO (Either ByteString String)

insert one or more rows

ping :: GenHaxl u w ByteString

exec :: HttpEnvironment a => String -> Env a w -> IO (Either ByteString String)

actual function used by user to perform fetching command

insertFromFile :: HttpEnvironment a => String -> Format -> FilePath -> Env a w -> IO (Either ByteString String)

insert data from

Connection

defaultHttpClient :: MonadIO m => m (Env HttpConnection w)

httpClient :: MonadIO m => String -> String -> m (Env HttpConnection w)

defaultHttpPool :: MonadIO m => Int -> NominalDiffTime -> Int -> m (Env (Pool HttpConnection) w)