ClickHouseDriver.IO.BufferedReader

Description

Tools to analyze protocol and deserialize data sent from server. This module is for internal use only.

Documentation

readBinaryStrWithLength :: Int -> Reader ByteString

readVarInt'

Arguments

:: Buffer
buffer to be read

-> IO (Word, Buffer)
(the word read from buffer, the buffer after reading)

readBinaryStr'

Arguments

:: Buffer
Buffer to be read

-> IO (ByteString, Buffer)
(the string read from Buffer, the buffer after reading)

read binary string from buffer. It first read the integer(n) in front of the desired string, then it read n bytes to capture the whole string.

readBinaryStr :: Reader ByteString

readVarInt :: Reader Word

readBinaryInt8 :: Reader Int8

readBinaryInt16 :: Reader Int16

readBinaryInt64 :: Reader Int64

readBinaryInt32 :: Reader Int32

readBinaryUInt8 :: Reader Word8

readBinaryUInt128 :: Reader Word128

readBinaryUInt64 :: Reader Word64

readBinaryUInt32 :: Reader Word32

readBinaryUInt16 :: Reader Word16

type Reader a = StateT Buffer IO a

data Buffer

Buffer is for receiving data from TCP stream. Whenever all bytes are read, it automatically refill from the stream.

Constructors

Buffer

Fields

bufSize :: !Int

bytesData :: ByteString

socket :: Maybe Socket

createBuffer :: Int -> Socket -> IO Buffer

create buffer with size and socket.

refill :: Buffer -> IO Buffer

refill buffer from stream